Example 2.B.4 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup(p-56)

Exam2.B.4 is used to illustrate one way treatment design with Binomial observations.

References

Stroup, W. W. (2012). Generalized Linear Mixed Models: Modern Concepts, Methods and Applications. CRC Press.

See also

Examples

#-----------------------------------------------------------------------------------
## logit Model  discussed in Example 2.B.2 using DataExam2.B.4
## Default link is logit
## using fmaily=binomial gives warning message of no-integer successes
#-----------------------------------------------------------------------------------
data(DataExam2.B.4)
DataExam2.B.4$trt <- factor(x =  DataExam2.B.4$trt)
Exam2.B.4glm <-
  glm(
         formula    =  Yij/Nij~trt
       , family     =  quasibinomial(link = "probit")
       , data       =  DataExam2.B.4
       , weights    =  NULL
    #  , subset
    #  , na.action
       , start      =  NULL
    #  , etastart
    #  , mustart
    #  , offset
    #  , control    =  list(...)
    #  , model      =  TRUE
       , method     =  "glm.fit"
    #  , x          =  FALSE
    #  , y          =  TRUE
       , contrasts  =  NULL
    #  , ...
  )
summary(Exam2.B.4glm)

#> 
#> Call:
#> glm(formula = Yij/Nij ~ trt, family = quasibinomial(link = "probit"), 
#>     data = DataExam2.B.4, weights = NULL, start = NULL, method = "glm.fit", 
#>     contrasts = NULL)
#> 
#> Deviance Residuals: 
#>       1        2        3        4        5        6  
#> -0.1660   0.1542  -0.3896   0.3896   0.2044  -0.2007  
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)
#> (Intercept)  -0.7729     0.3737  -2.068    0.130
#> trt2          0.7729     0.5017   1.541    0.221
#> trt3          0.9408     0.5029   1.871    0.158
#> 
#> (Dispersion parameter for quasibinomial family taken to be 0.1426701)
#> 
#>     Null deviance: 1.01671  on 5  degrees of freedom
#> Residual deviance: 0.43701  on 3  degrees of freedom
#> AIC: NA
#> 
#> Number of Fisher Scoring iterations: 4
#>